requirements.txt
requests
python-dotenv
qrcode[pil]
pyzbar
flask
flask_cors